rename issue test to reference PR number
